Move fixture file into fixture folder
